:art:  split code into domains
